requirements/local.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/local.in
#
--no-binary psycopg2

alabaster==0.7.12         # via sphinx
amqp==2.5.1
appdirs==1.4.3            # via black
argon2-cffi==19.1.0
astroid==2.2.5            # via pylint, pylint-celery
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, packaging, pytest
babel==2.7.0              # via sphinx
backcall==0.1.0           # via ipython
billiard==3.6.1.0
black==19.3b0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==7.0                # via black
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
decorator==4.4.0          # via ipython, traitlets
defusedxml==0.6.0
django-appconf==1.0.3
django-autoslug==1.9.6
django-celery-beat==1.5.0
django-choices==1.7.0
django-compressor==2.3
django-coverage-plugin==1.6.0
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-model-utils==3.2.0
django-redis==4.10.0
django-timezone-field==3.0
django==2.2.5
djangorestframework==3.10.3
docutils==0.15.2          # via sphinx
ecdsa==0.13.2
entrypoints==0.3          # via flake8
factory-boy==2.12.0
faker==2.0.1              # via factory-boy
flake8==3.7.8
future==0.17.1
idna==2.8
imagesize==1.1.0          # via sphinx
importlib-metadata==0.21
ipdb==0.12.2
ipython-genutils==0.2.0   # via traitlets
ipython==7.8.0            # via ipdb
isort==4.3.21             # via pylint
itypes==1.1.0
jedi==0.15.1              # via ipython
jinja2==2.10.1
kombu==4.6.4
lazy-object-proxy==1.4.2  # via astroid
markupsafe==1.1.1
mccabe==0.6.1             # via flake8, pylint
more-itertools==7.2.0
mypy-extensions==0.4.1    # via mypy
mypy==0.720
oauthlib==3.1.0
packaging==19.1           # via pytest, pytest-sugar, sphinx
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.1.0
pluggy==0.12.0            # via pytest
prompt-toolkit==2.0.9     # via ipython
psycopg2==2.8.3
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pyasn1==0.4.7
pycodestyle==2.5.0        # via flake8
pycparser==2.19
pyflakes==2.1.1           # via flake8
pygments==2.4.2           # via ipython, sphinx
pyjwt==1.7.1
pylint-celery==0.3
pylint-django==2.0.11
pylint-plugin-utils==0.5  # via pylint-celery, pylint-django
pylint==2.3.1             # via pylint-celery, pylint-django, pylint-plugin-utils
pyparsing==2.4.2          # via packaging
pytest-django==3.5.1
pytest-sugar==0.9.2
pytest==5.1.2
python-crontab==2.3.8
python-dateutil==2.8.0
python-jose==3.0.1
python-redis-lock[django]==3.3.1
python-slugify==3.0.3
python3-openid==3.1.0
pytz==2019.2
rcssmin==1.0.6
redis==3.3.8
requests-oauthlib==1.2.0
requests==2.22.0
rest-social-auth==2.2.0
rjsmin==1.1.0
rsa==4.0
rules==2.1
six==1.12.0
snowballstemmer==1.9.1    # via sphinx
social-auth-app-django==3.1.0
social-auth-core[openidconnect]==3.2.0
sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1  # via sphinx
sphinxcontrib-devhelp==1.0.1  # via sphinx
sphinxcontrib-htmlhelp==1.0.2  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.2  # via sphinx
sphinxcontrib-serializinghtml==1.1.3  # via sphinx
sqlparse==0.3.0
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.2
toml==0.10.0              # via black
traitlets==4.3.2          # via ipython
typed-ast==1.4.0          # via astroid, mypy
typing-extensions==3.7.4  # via mypy
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7            # via prompt-toolkit, pytest
werkzeug==0.14.1
wrapt==1.11.2             # via astroid
zipp==0.6.0
